feat: decouple from takumi

+++ api_story_test.go
@@ -29,7 +29,7 @@ func TestStoryService_GetStoryCategories(t *testing.T) {
 	}))
 
 	categories, _, err := client.StoryService.GetStoryCategories(ctx, &GetStoryCategoriesRequest{
-		WorkspaceID: Ptr(11112222),
+		WorkspaceID: Ptr(int64(11112222)),
 		ID:          NewMulti[int64](1111111111111, 1111111111112),
 		Name:        Ptr("test name"),
 		Description: Ptr("test description"),
